Feat: add map state to props for services
