+++ app/src/containers/Services/MainServices/MainServices.js
@@ -4,39 +4,6 @@ import {
   Row
 } from 'react-foundation';
 
-const services = [
-  {
-    id: 0,
-    title: 'Progressive Web',
-    description: 'With years of experience and expertise building progressive web software built for the Internet of Things. I have built software in use in appliances and smart products across the world and can help you to build out your product using the lateset and greatest web technologies available today.'
-  },
-  {
-    id: 1,
-    title: 'Mobile',
-    description: 'I focus on creating web software with a mobile-first approach using responsive web technologies and progressive JavaScript frameworks. Utilizing Swift for native iOS applications and React Native, I can build you pixel perfect mobile applications.'
-  },
-  {
-    id: 2,
-    title: 'Big Data / Machine Learning',
-    description: 'With experience in big data and Deep Learning, I can help make your software smarter.'
-  },
-  {
-    id: 3,
-    title: 'Data Architecture',
-    description: 'You don\'t need to be Google to employ an intelligent cloud computing system. With nearly 8 years experience architecting scalable data systems, I can help you to implement the latest and greatest real-time distributed server technologies, including Docker, MongoDB, SQL, Puppet, Chef, et. al.'
-  },
-  {
-    id: 4,
-    title: 'Agile Project Management',
-    description: 'Agile software development helps you to get software into the hands of your end users faster. Utilizing Scrum and other Agile methodologies, I can help you to keep your project moving on schedule.'
-  },
-  {
-    id: 5,
-    title: 'Pro Bono Consulting',
-    description: 'Through the <a href="https://hacksmiths.io">Hacksmiths.io</a> platform, I offer Pro Bono software development and consulting services to charitable and philanthrophic organizations. Please contact me if your charitable organization can use the help of a talented group of professional software developers!'
-  }
-];
-
 const MainServiceItem = ({
   item
 }) => (
