+++ app/src/containers/ServicesPage/ServicesPage.js
@@ -4,7 +4,8 @@ import {
   ServicesSectionTwo,
   ServicesSectionOne,
   VideoSeries,
-  Team
+  Team,
+  MainServices
 } from 'containers';
 import {
   Column,
@@ -12,6 +13,7 @@ import {
   Button
 } from 'react-foundation';
 import './ServicesPage.scss';
+import { connect } from 'react-redux';
 
 const ServicesPage = props => (
   <div className="services-page-wrapper">
@@ -28,10 +30,9 @@ const ServicesPage = props => (
       >
       <div className="services-intro">
         <p>Included on this page are some of the services that I offer as an independant consultant.</p>
-        <p>In some cases, I may try to connect you to one of my colleagues if you are seeking a specific set of skills that I don't have.</p>
-        <p>Head on over to <a href="http://hacksmiths.io">Hacksmiths.io</a> for more information about the Hacksmiths Platform and what we do.</p>
       </div>
       </Column>
+      <MainServices services={props.services} />
       <Column
         large={8}
         medium={10}
@@ -54,4 +55,12 @@ const ServicesPage = props => (
   </div>
 );
 
-export default ServicesPage;
+ServicesPage.propTypes = {
+  services: PropTypes.array.isRequired
+};
+
+const mapStateToProps = (state) => ({
+  services: state.services
+});
+
+export default connect(mapStateToProps)(ServicesPage);
